Better dir structure, README.md
