+++ hbase/app/src/main/resources/hbase-site.xml
@@ -1,0 +1,14 @@
+<configuration>
+  <property>
+    <name>hbase.zookeeper.quorum</name>
+    <value>zoo</value>
+  </property>
+  <property>
+    <name>hbase.zookeeper.property.clientPort</name>
+    <value>2181</value>
+  </property>
+  <property>
+    <name>hbase.client.scanner.caching</name>
+    <value>1000</value>
+  </property>
+</configuration>
