+++ hbase/docker/Dockerfile-datanode-snappy
@@ -1,0 +1,7 @@
+FROM bde2020/hadoop-datanode:latest
+
+MAINTAINER Aleksandar Mastilovic <amastilovic@sightmachine.com>
+
+COPY natives.tar.gz /natives.tar.gz
+
+RUN cd / && tar -xzf /natives.tar.gz -C / && rm -f /natives.tar.gz
