+++ hbase/docker/conf/backup-masters
@@ -1,0 +1,1 @@
+hbase-docker_hbase-regionserver_1
