+++ hbase/docker/conf/regionservers
@@ -1,0 +1,3 @@
+hbase-docker_hbase-regionserver_1
+hbase-docker_hbase-regionserver_2
+
