+++ hbase/docker/hadoop.env
@@ -1,0 +1,50 @@
+CORE_CONF_fs_defaultFS=hdfs://namenode:9000
+CORE_CONF_hadoop_http_staticuser_user=root
+CORE_CONF_hadoop_proxyuser_hue_hosts=*
+CORE_CONF_hadoop_proxyuser_hue_groups=*
+CORE_CONF_io_compression_codecs=org.apache.hadoop.io.compress.SnappyCodec
+
+HDFS_CONF_dfs_webhdfs_enabled=false
+HDFS_CONF_dfs_permissions_enabled=false
+HDFS_CONF_dfs_namenode_datanode_registration_ip___hostname___check=false
+HDFS_CONF_file_replication=1
+HDFS_CONF_dfs_name_dir=/hadoop/dfs/name
+HDFS_CONF_dfs_data_dir=/hadoop/dfs/data
+HDFS_CONF_dfs_replication=1
+HDFS_CONF_dfs_namenode_maintenance_replication_min=1
+# YARN_CONF_yarn_log___aggregation___enable=true
+# YARN_CONF_yarn_log_server_url=http://historyserver:8188/applicationhistory/logs/
+YARN_CONF_yarn_resourcemanager_recovery_enabled=true
+YARN_CONF_yarn_resourcemanager_store_class=org.apache.hadoop.yarn.server.resourcemanager.recovery.FileSystemRMStateStore
+YARN_CONF_yarn_resourcemanager_scheduler_class=org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacityScheduler
+YARN_CONF_yarn_scheduler_capacity_root_default_maximum___allocation___mb=1024
+YARN_CONF_yarn_scheduler_capacity_root_default_maximum___allocation___vcores=1
+YARN_CONF_yarn_resourcemanager_fs_state___store_uri=/rmstate
+YARN_CONF_yarn_resourcemanager_system___metrics___publisher_enabled=true
+YARN_CONF_yarn_resourcemanager_hostname=resourcemanager
+YARN_CONF_yarn_resourcemanager_address=resourcemanager:8032
+YARN_CONF_yarn_resourcemanager_scheduler_address=resourcemanager:8030
+YARN_CONF_yarn_resourcemanager_resource__tracker_address=resourcemanager:8031
+# YARN_CONF_yarn_timeline___service_enabled=false
+# YARN_CONF_yarn_timeline___service_generic___application___history_enabled=false
+# YARN_CONF_yarn_timeline___service_hostname=historyserver
+YARN_CONF_mapreduce_map_output_compress=true
+YARN_CONF_mapred_map_output_compress_codec=org.apache.hadoop.io.compress.SnappyCodec
+YARN_CONF_yarn_nodemanager_resource_memory___mb=1024
+YARN_CONF_yarn_nodemanager_resource_cpu___vcores=1
+YARN_CONF_yarn_nodemanager_disk___health___checker_max___disk___utilization___per___disk___percentage=98.5
+YARN_CONF_yarn_nodemanager_remote___app___log___dir=/app-logs
+YARN_CONF_yarn_nodemanager_aux___services=mapreduce_shuffle
+
+MAPRED_CONF_mapreduce_framework_name=yarn
+MAPRED_CONF_mapred_child_java_opts=-Xmx550m
+MAPRED_CONF_mapreduce_map_memory_mb=512
+MAPRED_CONF_mapreduce_reduce_memory_mb=512
+MAPRED_CONF_mapreduce_map_java_opts=-Xmx550m
+MAPRED_CONF_mapreduce_reduce_java_opts=-Xmx1100m
+# original:
+# MAPRED_CONF_mapreduce_map_memory_mb=4096
+# MAPRED_CONF_mapreduce_reduce_memory_mb=8192
+# MAPRED_CONF_mapreduce_map_java_opts=-Xmx3072m
+# MAPRED_CONF_mapreduce_reduce_java_opts=-Xmx6144m
+
