+++ hbase/docker/hbase-distributed-local.env
@@ -1,0 +1,13 @@
+HBASE_CONF_hbase_rootdir=hdfs://namenode:9000/hbase
+HBASE_CONF_hbase_cluster_distributed=true
+HBASE_CONF_hbase_zookeeper_quorum=hbase-docker_zoo_1_1,hbase-docker_zoo_2_1
+HBASE_CONF_hbase_zookeeper_property_clientPort=2181
+# HBASE_CONF_hbase_zookeeper_property_dataDir=hdfs://namenode:9000/zookeeper
+HBASE_CONF_hbase_master=hbase-master:16000
+HBASE_CONF_hbase_master_hostname=hbase-master
+HBASE_CONF_hbase_master_port=16000
+HBASE_CONF_hbase_master_info_port=16010
+HBASE_CONF_hbase_regionserver_port=16020
+HBASE_CONF_hbase_regionserver_info_port=16030
+
+HBASE_MANAGES_ZK=false
