+++ hbase/docker/hbase-standalone.env
@@ -1,0 +1,13 @@
+HBASE_CONF_hbase_rootdir=file:///hbase-data/hbase
+HBASE_CONF_hbase_cluster_distributed=true
+HBASE_CONF_hbase_zookeeper_property_dataDir=/zookeeper-data
+HBASE_CONF_hbase_zookeeper_quorum=zoo
+
+HBASE_CONF_hbase_master_port=16000
+HBASE_CONF_hbase_master_info_port=16010
+HBASE_CONF_hbase_regionserver_port=16020
+HBASE_CONF_hbase_regionserver_info_port=16030
+HBASE_CONF_hbase_zookeeper_peerport=2888
+HBASE_CONF_hbase_zookeeper_leaderport=3888
+HBASE_CONF_hbase_zookeeper_property_clientPort=2181
+
